INS-2644: Move all contracts from insgorund to builtin
