--- application/contract/nodedomain/nodedomain.go
@@ -1,88 +1,0 @@
-//
-// Copyright 2019 Insolar Technologies GmbH
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-
-package nodedomain
-
-import (
-	"fmt"
-
-	"github.com/insolar/insolar/application/proxy/noderecord"
-	"github.com/insolar/insolar/application/proxy/rootdomain"
-	"github.com/insolar/insolar/insolar"
-	"github.com/insolar/insolar/logicrunner/goplugin/foundation"
-)
-
-// NodeDomain holds noderecords
-type NodeDomain struct {
-	foundation.BaseContract
-
-	NodeIndexPublicKey map[string]string
-}
-
-// NewNodeDomain create new NodeDomain
-func NewNodeDomain() (*NodeDomain, error) {
-	return &NodeDomain{
-		NodeIndexPublicKey: make(map[string]string),
-	}, nil
-}
-
-func (nd *NodeDomain) getNodeRecord(ref insolar.Reference) *noderecord.NodeRecord {
-	return noderecord.GetObject(ref)
-}
-
-// RegisterNode registers node in system
-func (nd *NodeDomain) RegisterNode(publicKey string, role string) (string, error) {
-
-	root, err := rootdomain.GetObject(*nd.GetContext().Parent).GetRootMemberRef()
-	if err != nil {
-		return "", fmt.Errorf("failed to get root member reference: %s", err.Error())
-	}
-	if *nd.GetContext().Caller != *root {
-		return "", fmt.Errorf("only root member can register node")
-	}
-
-	newNode := noderecord.NewNodeRecord(publicKey, role)
-	node, err := newNode.AsChild(nd.GetReference())
-	if err != nil {
-		return "", fmt.Errorf("failed to save as child: %s", err.Error())
-	}
-
-	newNodeRef := node.GetReference().String()
-	nd.NodeIndexPublicKey[publicKey] = newNodeRef
-
-	return newNodeRef, err
-}
-
-// GetNodeRefByPublicKey returns node ref
-func (nd *NodeDomain) GetNodeRefByPublicKey(publicKey string) (string, error) {
-	nodeRef, ok := nd.NodeIndexPublicKey[publicKey]
-	if !ok {
-		return nodeRef, fmt.Errorf("network node not found by public key: %s", publicKey)
-	}
-	return nodeRef, nil
-}
-
-// RemoveNode deletes node from registry
-func (nd *NodeDomain) RemoveNode(nodeRef insolar.Reference) error {
-	node := nd.getNodeRecord(nodeRef)
-	nodePK, err := node.GetPublicKey()
-	if err != nil {
-		return fmt.Errorf("failed to find network node by public key: %s", nodePK)
-	}
-
-	delete(nd.NodeIndexPublicKey, nodePK)
-	return node.Destroy()
-}
