--- application/proxy/costcenter/costcenter.go
@@ -1,494 +1,0 @@
-//
-// Copyright 2019 Insolar Technologies GmbH
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-
-package costcenter
-
-import (
-	"github.com/insolar/insolar/insolar"
-	"github.com/insolar/insolar/logicrunner/common"
-	"github.com/insolar/insolar/logicrunner/goplugin/foundation"
-)
-
-// PrototypeReference to prototype of this contract
-// error checking hides in generator
-var PrototypeReference, _ = insolar.NewReferenceFromBase58("111158WTbUui5i7T1vetvbnCAU6JyWf2hm7NbPLWqc.11111111111111111111111111111111")
-
-// CostCenter holds proxy type
-type CostCenter struct {
-	Reference insolar.Reference
-	Prototype insolar.Reference
-	Code      insolar.Reference
-}
-
-// ContractConstructorHolder holds logic with object construction
-type ContractConstructorHolder struct {
-	constructorName string
-	argsSerialized  []byte
-}
-
-// AsChild saves object as child
-func (r *ContractConstructorHolder) AsChild(objRef insolar.Reference) (*CostCenter, error) {
-	ref, err := common.CurrentProxyCtx.SaveAsChild(objRef, *PrototypeReference, r.constructorName, r.argsSerialized)
-	if err != nil {
-		return nil, err
-	}
-	return &CostCenter{Reference: ref}, nil
-}
-
-// AsDelegate saves object as delegate
-func (r *ContractConstructorHolder) AsDelegate(objRef insolar.Reference) (*CostCenter, error) {
-	ref, err := common.CurrentProxyCtx.SaveAsDelegate(objRef, *PrototypeReference, r.constructorName, r.argsSerialized)
-	if err != nil {
-		return nil, err
-	}
-	return &CostCenter{Reference: ref}, nil
-}
-
-// GetObject returns proxy object
-func GetObject(ref insolar.Reference) (r *CostCenter) {
-	return &CostCenter{Reference: ref}
-}
-
-// GetPrototype returns reference to the prototype
-func GetPrototype() insolar.Reference {
-	return *PrototypeReference
-}
-
-// GetImplementationFrom returns proxy to delegate of given type
-func GetImplementationFrom(object insolar.Reference) (*CostCenter, error) {
-	ref, err := common.CurrentProxyCtx.GetDelegate(object, *PrototypeReference)
-	if err != nil {
-		return nil, err
-	}
-	return GetObject(ref), nil
-}
-
-// New is constructor
-func New(commissionWallet insolar.Reference, currentTariff insolar.Reference) *ContractConstructorHolder {
-	var args [2]interface{}
-	args[0] = commissionWallet
-	args[1] = currentTariff
-
-	var argsSerialized []byte
-	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
-	if err != nil {
-		panic(err)
-	}
-
-	return &ContractConstructorHolder{constructorName: "New", argsSerialized: argsSerialized}
-}
-
-// GetReference returns reference of the object
-func (r *CostCenter) GetReference() insolar.Reference {
-	return r.Reference
-}
-
-// GetPrototype returns reference to the code
-func (r *CostCenter) GetPrototype() (insolar.Reference, error) {
-	if r.Prototype.IsEmpty() {
-		ret := [2]interface{}{}
-		var ret0 insolar.Reference
-		ret[0] = &ret0
-		var ret1 *foundation.Error
-		ret[1] = &ret1
-
-		res, err := common.CurrentProxyCtx.RouteCall(r.Reference, true, false, "GetPrototype", make([]byte, 0), *PrototypeReference)
-		if err != nil {
-			return ret0, err
-		}
-
-		err = common.CurrentProxyCtx.Deserialize(res, &ret)
-		if err != nil {
-			return ret0, err
-		}
-
-		if ret1 != nil {
-			return ret0, ret1
-		}
-
-		r.Prototype = ret0
-	}
-
-	return r.Prototype, nil
-
-}
-
-// GetCode returns reference to the code
-func (r *CostCenter) GetCode() (insolar.Reference, error) {
-	if r.Code.IsEmpty() {
-		ret := [2]interface{}{}
-		var ret0 insolar.Reference
-		ret[0] = &ret0
-		var ret1 *foundation.Error
-		ret[1] = &ret1
-
-		res, err := common.CurrentProxyCtx.RouteCall(r.Reference, true, false, "GetCode", make([]byte, 0), *PrototypeReference)
-		if err != nil {
-			return ret0, err
-		}
-
-		err = common.CurrentProxyCtx.Deserialize(res, &ret)
-		if err != nil {
-			return ret0, err
-		}
-
-		if ret1 != nil {
-			return ret0, ret1
-		}
-
-		r.Code = ret0
-	}
-
-	return r.Code, nil
-}
-
-// SetTariffs is proxy generated method
-func (r *CostCenter) SetTariffs(tariffs []insolar.Reference) error {
-	var args [1]interface{}
-	args[0] = tariffs
-
-	var argsSerialized []byte
-
-	ret := [1]interface{}{}
-	var ret0 *foundation.Error
-	ret[0] = &ret0
-
-	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
-	if err != nil {
-		return err
-	}
-
-	res, err := common.CurrentProxyCtx.RouteCall(r.Reference, true, false, "SetTariffs", argsSerialized, *PrototypeReference)
-	if err != nil {
-		return err
-	}
-
-	err = common.CurrentProxyCtx.Deserialize(res, &ret)
-	if err != nil {
-		return err
-	}
-
-	if ret0 != nil {
-		return ret0
-	}
-	return nil
-}
-
-// SetTariffsNoWait is proxy generated method
-func (r *CostCenter) SetTariffsNoWait(tariffs []insolar.Reference) error {
-	var args [1]interface{}
-	args[0] = tariffs
-
-	var argsSerialized []byte
-
-	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
-	if err != nil {
-		return err
-	}
-
-	_, err = common.CurrentProxyCtx.RouteCall(r.Reference, false, false, "SetTariffs", argsSerialized, *PrototypeReference)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// SetTariffsAsImmutable is proxy generated method
-func (r *CostCenter) SetTariffsAsImmutable(tariffs []insolar.Reference) error {
-	var args [1]interface{}
-	args[0] = tariffs
-
-	var argsSerialized []byte
-
-	ret := [1]interface{}{}
-	var ret0 *foundation.Error
-	ret[0] = &ret0
-
-	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
-	if err != nil {
-		return err
-	}
-
-	res, err := common.CurrentProxyCtx.RouteCall(r.Reference, true, true, "SetTariffs", argsSerialized, *PrototypeReference)
-	if err != nil {
-		return err
-	}
-
-	err = common.CurrentProxyCtx.Deserialize(res, &ret)
-	if err != nil {
-		return err
-	}
-
-	if ret0 != nil {
-		return ret0
-	}
-	return nil
-}
-
-// GetTariffs is proxy generated method
-func (r *CostCenter) GetTariffs() ([]insolar.Reference, error) {
-	var args [0]interface{}
-
-	var argsSerialized []byte
-
-	ret := [2]interface{}{}
-	var ret0 []insolar.Reference
-	ret[0] = &ret0
-	var ret1 *foundation.Error
-	ret[1] = &ret1
-
-	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
-	if err != nil {
-		return ret0, err
-	}
-
-	res, err := common.CurrentProxyCtx.RouteCall(r.Reference, true, false, "GetTariffs", argsSerialized, *PrototypeReference)
-	if err != nil {
-		return ret0, err
-	}
-
-	err = common.CurrentProxyCtx.Deserialize(res, &ret)
-	if err != nil {
-		return ret0, err
-	}
-
-	if ret1 != nil {
-		return ret0, ret1
-	}
-	return ret0, nil
-}
-
-// GetTariffsNoWait is proxy generated method
-func (r *CostCenter) GetTariffsNoWait() error {
-	var args [0]interface{}
-
-	var argsSerialized []byte
-
-	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
-	if err != nil {
-		return err
-	}
-
-	_, err = common.CurrentProxyCtx.RouteCall(r.Reference, false, false, "GetTariffs", argsSerialized, *PrototypeReference)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// GetTariffsAsImmutable is proxy generated method
-func (r *CostCenter) GetTariffsAsImmutable() ([]insolar.Reference, error) {
-	var args [0]interface{}
-
-	var argsSerialized []byte
-
-	ret := [2]interface{}{}
-	var ret0 []insolar.Reference
-	ret[0] = &ret0
-	var ret1 *foundation.Error
-	ret[1] = &ret1
-
-	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
-	if err != nil {
-		return ret0, err
-	}
-
-	res, err := common.CurrentProxyCtx.RouteCall(r.Reference, true, true, "GetTariffs", argsSerialized, *PrototypeReference)
-	if err != nil {
-		return ret0, err
-	}
-
-	err = common.CurrentProxyCtx.Deserialize(res, &ret)
-	if err != nil {
-		return ret0, err
-	}
-
-	if ret1 != nil {
-		return ret0, ret1
-	}
-	return ret0, nil
-}
-
-// SetCurrentTariff is proxy generated method
-func (r *CostCenter) SetCurrentTariff(currentTariff insolar.Reference) error {
-	var args [1]interface{}
-	args[0] = currentTariff
-
-	var argsSerialized []byte
-
-	ret := [1]interface{}{}
-	var ret0 *foundation.Error
-	ret[0] = &ret0
-
-	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
-	if err != nil {
-		return err
-	}
-
-	res, err := common.CurrentProxyCtx.RouteCall(r.Reference, true, false, "SetCurrentTariff", argsSerialized, *PrototypeReference)
-	if err != nil {
-		return err
-	}
-
-	err = common.CurrentProxyCtx.Deserialize(res, &ret)
-	if err != nil {
-		return err
-	}
-
-	if ret0 != nil {
-		return ret0
-	}
-	return nil
-}
-
-// SetCurrentTariffNoWait is proxy generated method
-func (r *CostCenter) SetCurrentTariffNoWait(currentTariff insolar.Reference) error {
-	var args [1]interface{}
-	args[0] = currentTariff
-
-	var argsSerialized []byte
-
-	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
-	if err != nil {
-		return err
-	}
-
-	_, err = common.CurrentProxyCtx.RouteCall(r.Reference, false, false, "SetCurrentTariff", argsSerialized, *PrototypeReference)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// SetCurrentTariffAsImmutable is proxy generated method
-func (r *CostCenter) SetCurrentTariffAsImmutable(currentTariff insolar.Reference) error {
-	var args [1]interface{}
-	args[0] = currentTariff
-
-	var argsSerialized []byte
-
-	ret := [1]interface{}{}
-	var ret0 *foundation.Error
-	ret[0] = &ret0
-
-	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
-	if err != nil {
-		return err
-	}
-
-	res, err := common.CurrentProxyCtx.RouteCall(r.Reference, true, true, "SetCurrentTariff", argsSerialized, *PrototypeReference)
-	if err != nil {
-		return err
-	}
-
-	err = common.CurrentProxyCtx.Deserialize(res, &ret)
-	if err != nil {
-		return err
-	}
-
-	if ret0 != nil {
-		return ret0
-	}
-	return nil
-}
-
-// GetCurrentTariff is proxy generated method
-func (r *CostCenter) GetCurrentTariff() (insolar.Reference, error) {
-	var args [0]interface{}
-
-	var argsSerialized []byte
-
-	ret := [2]interface{}{}
-	var ret0 insolar.Reference
-	ret[0] = &ret0
-	var ret1 *foundation.Error
-	ret[1] = &ret1
-
-	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
-	if err != nil {
-		return ret0, err
-	}
-
-	res, err := common.CurrentProxyCtx.RouteCall(r.Reference, true, false, "GetCurrentTariff", argsSerialized, *PrototypeReference)
-	if err != nil {
-		return ret0, err
-	}
-
-	err = common.CurrentProxyCtx.Deserialize(res, &ret)
-	if err != nil {
-		return ret0, err
-	}
-
-	if ret1 != nil {
-		return ret0, ret1
-	}
-	return ret0, nil
-}
-
-// GetCurrentTariffNoWait is proxy generated method
-func (r *CostCenter) GetCurrentTariffNoWait() error {
-	var args [0]interface{}
-
-	var argsSerialized []byte
-
-	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
-	if err != nil {
-		return err
-	}
-
-	_, err = common.CurrentProxyCtx.RouteCall(r.Reference, false, false, "GetCurrentTariff", argsSerialized, *PrototypeReference)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// GetCurrentTariffAsImmutable is proxy generated method
-func (r *CostCenter) GetCurrentTariffAsImmutable() (insolar.Reference, error) {
-	var args [0]interface{}
-
-	var argsSerialized []byte
-
-	ret := [2]interface{}{}
-	var ret0 insolar.Reference
-	ret[0] = &ret0
-	var ret1 *foundation.Error
-	ret[1] = &ret1
-
-	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
-	if err != nil {
-		return ret0, err
-	}
-
-	res, err := common.CurrentProxyCtx.RouteCall(r.Reference, true, true, "GetCurrentTariff", argsSerialized, *PrototypeReference)
-	if err != nil {
-		return ret0, err
-	}
-
-	err = common.CurrentProxyCtx.Deserialize(res, &ret)
-	if err != nil {
-		return ret0, err
-	}
-
-	if ret1 != nil {
-		return ret0, ret1
-	}
-	return ret0, nil
-}
