+++ logicrunner/builtin/proxy/tariff/tariff.go
@@ -1,0 +1,243 @@
+//
+// Copyright 2019 Insolar Technologies GmbH
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+package tariff
+
+import (
+	"github.com/insolar/insolar/insolar"
+	"github.com/insolar/insolar/logicrunner/common"
+	"github.com/insolar/insolar/logicrunner/goplugin/foundation"
+)
+
+// PrototypeReference to prototype of this contract
+// error checking hides in generator
+var PrototypeReference, _ = insolar.NewReferenceFromBase58("111A7sBiaB3WxTAEa6X8xLcNGetQUpf3BsYZenTuGUW.11111111111111111111111111111111")
+
+// Tariff holds proxy type
+type Tariff struct {
+	Reference insolar.Reference
+	Prototype insolar.Reference
+	Code      insolar.Reference
+}
+
+// ContractConstructorHolder holds logic with object construction
+type ContractConstructorHolder struct {
+	constructorName string
+	argsSerialized  []byte
+}
+
+// AsChild saves object as child
+func (r *ContractConstructorHolder) AsChild(objRef insolar.Reference) (*Tariff, error) {
+	ref, err := common.CurrentProxyCtx.SaveAsChild(objRef, *PrototypeReference, r.constructorName, r.argsSerialized)
+	if err != nil {
+		return nil, err
+	}
+	return &Tariff{Reference: ref}, nil
+}
+
+// AsDelegate saves object as delegate
+func (r *ContractConstructorHolder) AsDelegate(objRef insolar.Reference) (*Tariff, error) {
+	ref, err := common.CurrentProxyCtx.SaveAsDelegate(objRef, *PrototypeReference, r.constructorName, r.argsSerialized)
+	if err != nil {
+		return nil, err
+	}
+	return &Tariff{Reference: ref}, nil
+}
+
+// GetObject returns proxy object
+func GetObject(ref insolar.Reference) (r *Tariff) {
+	return &Tariff{Reference: ref}
+}
+
+// GetPrototype returns reference to the prototype
+func GetPrototype() insolar.Reference {
+	return *PrototypeReference
+}
+
+// GetImplementationFrom returns proxy to delegate of given type
+func GetImplementationFrom(object insolar.Reference) (*Tariff, error) {
+	ref, err := common.CurrentProxyCtx.GetDelegate(object, *PrototypeReference)
+	if err != nil {
+		return nil, err
+	}
+	return GetObject(ref), nil
+}
+
+// New is constructor
+func New(commissionRate string) *ContractConstructorHolder {
+	var args [1]interface{}
+	args[0] = commissionRate
+
+	var argsSerialized []byte
+	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
+	if err != nil {
+		panic(err)
+	}
+
+	return &ContractConstructorHolder{constructorName: "New", argsSerialized: argsSerialized}
+}
+
+// GetReference returns reference of the object
+func (r *Tariff) GetReference() insolar.Reference {
+	return r.Reference
+}
+
+// GetPrototype returns reference to the code
+func (r *Tariff) GetPrototype() (insolar.Reference, error) {
+	if r.Prototype.IsEmpty() {
+		ret := [2]interface{}{}
+		var ret0 insolar.Reference
+		ret[0] = &ret0
+		var ret1 *foundation.Error
+		ret[1] = &ret1
+
+		res, err := common.CurrentProxyCtx.RouteCall(r.Reference, true, false, "GetPrototype", make([]byte, 0), *PrototypeReference)
+		if err != nil {
+			return ret0, err
+		}
+
+		err = common.CurrentProxyCtx.Deserialize(res, &ret)
+		if err != nil {
+			return ret0, err
+		}
+
+		if ret1 != nil {
+			return ret0, ret1
+		}
+
+		r.Prototype = ret0
+	}
+
+	return r.Prototype, nil
+
+}
+
+// GetCode returns reference to the code
+func (r *Tariff) GetCode() (insolar.Reference, error) {
+	if r.Code.IsEmpty() {
+		ret := [2]interface{}{}
+		var ret0 insolar.Reference
+		ret[0] = &ret0
+		var ret1 *foundation.Error
+		ret[1] = &ret1
+
+		res, err := common.CurrentProxyCtx.RouteCall(r.Reference, true, false, "GetCode", make([]byte, 0), *PrototypeReference)
+		if err != nil {
+			return ret0, err
+		}
+
+		err = common.CurrentProxyCtx.Deserialize(res, &ret)
+		if err != nil {
+			return ret0, err
+		}
+
+		if ret1 != nil {
+			return ret0, ret1
+		}
+
+		r.Code = ret0
+	}
+
+	return r.Code, nil
+}
+
+// GetCommission is proxy generated method
+func (r *Tariff) GetCommission(amountStr string) (string, error) {
+	var args [1]interface{}
+	args[0] = amountStr
+
+	var argsSerialized []byte
+
+	ret := [2]interface{}{}
+	var ret0 string
+	ret[0] = &ret0
+	var ret1 *foundation.Error
+	ret[1] = &ret1
+
+	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
+	if err != nil {
+		return ret0, err
+	}
+
+	res, err := common.CurrentProxyCtx.RouteCall(r.Reference, true, false, "GetCommission", argsSerialized, *PrototypeReference)
+	if err != nil {
+		return ret0, err
+	}
+
+	err = common.CurrentProxyCtx.Deserialize(res, &ret)
+	if err != nil {
+		return ret0, err
+	}
+
+	if ret1 != nil {
+		return ret0, ret1
+	}
+	return ret0, nil
+}
+
+// GetCommissionNoWait is proxy generated method
+func (r *Tariff) GetCommissionNoWait(amountStr string) error {
+	var args [1]interface{}
+	args[0] = amountStr
+
+	var argsSerialized []byte
+
+	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
+	if err != nil {
+		return err
+	}
+
+	_, err = common.CurrentProxyCtx.RouteCall(r.Reference, false, false, "GetCommission", argsSerialized, *PrototypeReference)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+// GetCommissionAsImmutable is proxy generated method
+func (r *Tariff) GetCommissionAsImmutable(amountStr string) (string, error) {
+	var args [1]interface{}
+	args[0] = amountStr
+
+	var argsSerialized []byte
+
+	ret := [2]interface{}{}
+	var ret0 string
+	ret[0] = &ret0
+	var ret1 *foundation.Error
+	ret[1] = &ret1
+
+	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
+	if err != nil {
+		return ret0, err
+	}
+
+	res, err := common.CurrentProxyCtx.RouteCall(r.Reference, true, true, "GetCommission", argsSerialized, *PrototypeReference)
+	if err != nil {
+		return ret0, err
+	}
+
+	err = common.CurrentProxyCtx.Deserialize(res, &ret)
+	if err != nil {
+		return ret0, err
+	}
+
+	if ret1 != nil {
+		return ret0, ret1
+	}
+	return ret0, nil
+}
